fix(api,contrib): get run result without type filter
