+++ contrib/grpcplugins/action/junit/main.go
@@ -78,7 +78,8 @@ func (actPlugin *junitPlugin) perform(ctx context.Context, dirFS fs.FS, filePath
 		}
 		testFailed += nbFailed
 
-		if _, err := grpcplugins.UploadRunResult(ctx, &actPlugin.Common, &grpcplugins.IntegrationCache{}, runResultRequest, r.Result, openFiles[r.Path], sizes[r.Path], checksums[r.Path]); err != nil {
+		integrationCache := grpcplugins.NewIntegrationCache()
+		if _, err := grpcplugins.UploadRunResult(ctx, &actPlugin.Common, integrationCache, runResultRequest, r.Result, openFiles[r.Path], sizes[r.Path], checksums[r.Path]); err != nil {
 			_ = openFiles[r.Path].Close()
 			return err
 		}
