+++ engine/worker/pkg/workerruntime/types.go
@@ -37,9 +37,8 @@ type V2UpdateResultResponse struct {
 }
 
 type V2FilterRunResult struct {
-	Pattern              string
-	Type                 []sdk.V2WorkflowRunResultType
-	WithClearIntegration bool
+	Pattern string
+	Type    []sdk.V2WorkflowRunResultType
 }
 
 type WorkerConfig struct {
